correcting issues
