+++ examples/getting_started/force_tracking_parameter_optimization.py
@@ -13,7 +13,7 @@ from cocofest import (
 )
 
 # --- Building force to track ---#
-time, force = ExtractData.load_data("data/hand_cycling_force.bio")
+time, force = ExtractData.load_data("../data/hand_cycling_force.bio")
 force = force - force[0]
 force_tracking = [time, force]
