+++ examples/getting_started/frequency_optimization.py
