+++ examples/getting_started/muscle_model_identification.py
