+++ examples/getting_started/pulse_duration_optimization.py
